[PackageBuilder] Move FinderSanitizer here
