+++ packages/EasyCodingStandard/src/config/external-services.yml
@@ -24,3 +24,5 @@ services:
 
     PhpCsFixer\Tokenizer\Analyzer\NamespaceUsesAnalyzer: ~
     PhpCsFixer\Tokenizer\Analyzer\NamespacesAnalyzer: ~
+
+    Symplify\PackageBuilder\FileSystem\FinderSanitizer: ~
