--- packages/EasyCodingStandard/tests/Finder/FinderSanitizerSource/NestedDirectory/FileWithClass.php
@@ -1,5 +1,0 @@
-<?php declare(strict_types=1);
-
-final class FileWithClass
-{
-}
