+++ packages/PackageBuilder/tests/FileSystem/FinderSanitizerSource/NestedDirectory/FileWithClass.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+namespace Symplify\PackageBuilder\Tests\FileSystem\FinderSanitizerSource\NestedDirectory;
+
+final class FileWithClass
+{
+}
